Hands-on Coding/Hands-On-Projects-and-Case-Studies/Hands-On Projects/FastAPI-API-Solution-Init/requirements.txt
aerich==0.7.1
asyncpg==0.29.0
bcrypt==4.0.1
passlib==1.7.4
fastapi==0.90.0
tortoise-orm==0.19.2
uvicorn==0.20.0
jinja2==3.1.2